+++ libs/code-generator/src/lib/code-generator.spec.ts
@@ -1,10 +1,12 @@
 import * as fs from 'fs/promises'
-import { rest } from 'msw/'
-import { SetupServer, setupServer } from 'msw/node'
+import { HttpResponse, http } from 'msw'
+import type { SetupServer } from 'msw/node'
+import { setupServer } from 'msw/node'
 import * as path from 'path'
-import { FsTree, Tree } from './tree/tree'
-
-import { CodeGeneratorOptions, codeGenerator } from './code-generator'
+import type { CodeGeneratorOptions } from './code-generator'
+import { codeGenerator } from './code-generator'
+import type { Tree } from './tree/tree'
+import { FsTree } from './tree/tree'
 
 describe('code-generator', () => {
     let tree: Tree
@@ -15,9 +17,9 @@ describe('code-generator', () => {
     describe('templateType: dotnet-api', () => {
         beforeAll(() => {
             server = setupServer(
-                rest.get(
+                http.get(
                     'https://api.featureboard.dev/projects',
-                    async (req, res, ctx) => {
+                    async () => {
                         const file = await fs.readFile(
                             path.join(
                                 __dirname,
@@ -27,7 +29,7 @@ describe('code-generator', () => {
                                 encoding: 'utf8',
                             },
                         )
-                        return res(ctx.status(200), ctx.json(JSON.parse(file)))
+                        return HttpResponse.json(JSON.parse(file))
                     },
                 ),
             )