+++ libs/js-sdk/.eslintrc.cjs
@@ -1,0 +1,18 @@
+/** @type {import('@types/eslint').Linter.BaseConfig} */
+module.exports = {
+  extends: ['../../.eslintrc.json', 'prettier'],
+  overrides: [
+    {
+      parserOptions: {
+        tsconfigRootDir: __dirname,
+        project: ['./tsconfig.json'],
+      },
+      files: ['*.ts', '*.tsx'],
+      rules: {
+        '@typescript-eslint/no-unsafe-argument': 'off',
+        '@typescript-eslint/no-unsafe-return': 'off',
+      },
+    },
+  ],
+  ignorePatterns: ['!**/*', 'node_modules', 'out-tsc'],
+}