+++ libs/nx-plugin/src/generators/code-gen/generator.ts
@@ -1,7 +1,9 @@
 import { codeGenerator } from '@featureboard/code-generator'
-import { Tree, joinPathFragments, readProjectConfiguration } from '@nx/devkit'
-import * as prompts from 'prompts'
-import { CodeGenGeneratorSchema } from './schema'
+import type { Tree } from '@nx/devkit'
+import { joinPathFragments, readProjectConfiguration } from '@nx/devkit'
+import prompts from 'prompts'
+import { isDryRun } from '../../Shared/is-dry-run'
+import type { CodeGenGeneratorSchema } from './schema'
 
 export async function codeGenGenerator(
     tree: Tree,
@@ -30,10 +32,6 @@ export async function codeGenGenerator(
         featureBoardBearerToken,
         ...options,
     })
-}
-
-export function isDryRun(): boolean {
-    return process.argv.some((x) => x === '--dry-run')
 }
 
 export default codeGenGenerator