+++ libs/nx-plugin/src/index.ts
@@ -1,0 +1,2 @@
+export * from './executors/code-gen/executor'
+export * from './generators/code-gen/generator'